==> LLM/1_Prompt_Generation.ipynb <==
# Import packages
import pandas as pd

# Read in cleaned csv file
df = pd.read_csv("../Data/subject-info-cleaned.csv")
df.head()

# ECG Impressions
# Add ECG Impressions to dataframe
impressions = ["Ventricular Extrasystole", "Ventricular Tachycardia", "Non-sustained ventricular tachycardia (CH>10)", 
               "Paroxysmal supraventricular tachyarrhythmia", "Bradycardia"]

# Create ECG reports
class ECGReport:
    def __init__(self, ventricular_extrasystole, ventricular_tachycardia, non_sustained_ventricular_tachycardia, paroxysmal_supraventricular_tachyarrhythmia, bradycardia):
        self.ventricular_extrasystole = ventricular_extrasystole
        self.ventricular_tachycardia = ventricular_tachycardia
        self.non_sustained_ventricular_tachycardia = non_sustained_ventricular_tachycardia
        self.paroxysmal_supraventricular_tachyarrhythmia = paroxysmal_supraventricular_tachyarrhythmia
        self.bradycardia = bradycardia
        
    def interpret_ventricular_extrasystole(self):
        ventricular_extrasystole_dict = {
            0: "No",
            1: "Monomorphic",
            2: "Polymorphic",
            3: "Couplets"
        }
        return ventricular_extrasystole_dict.get(self.ventricular_extrasystole, "Unknown ventricular extrasystole code")
    
    def interpret_ventricular_tachycardia(self):
        ventricular_tachycardia_dict = {
            0: "No",
            1: "Non-sustained VT",
            2: "Sustained VT", 
            3: "Torsade de Points"
        }
        return ventricular_tachycardia_dict.get(self.ventricular_tachycardia, "Unknown ventricular tachycardia code")
    
    def interpret_non_sustained_ventricular_tachycardia(self):
        non_sustained_ventricular_tachycardia_dict = {
            0: "No",
            1: "Yes"
        }
        return non_sustained_ventricular_tachycardia_dict.get(self.non_sustained_ventricular_tachycardia, "Unknown non sustained ventricular tachycardia code")
    
    def interpret_paroxysmal_supraventricular_tachyarrhythmia(self):
        paroxysmal_supraventricular_tachyarrhythmia_dict = {
            0: "No", 
            1: "TPSV", 
            2: "Parosysmal AF", 
            3: "Paroxismal flutter", 
            4: "Others"
        }
        return paroxysmal_supraventricular_tachyarrhythmia_dict.get(self.paroxysmal_supraventricular_tachyarrhythmia, "Unknown paroxysmal supraventricular tachyarrhythmia code")
    
    def interpret_bradycardia(self):
        bradycardia_dict = {
            0: "No",
            1: "Sinus Node Dysfunction", 
            2: "First-degree Atrioventricular block (AVB)",
            3: "Second-degree AVB - type I",
            4: "Second-degree AVB - type II", 
            5: "Third-degree AVB", 
            6: "Paroxysmal AVB"
        }
        return bradycardia_dict.get(self.bradycardia, "Unknown bradycardia code")
    
    def generate_report(self):
        return f"""ECG Impression:
        - Ventricular Extrasystole: {self.interpret_ventricular_extrasystole()}
        - Ventricular Tachycardia: {self.interpret_ventricular_tachycardia()}
        - Non-sustained ventricular tachycardia (CH>10): {self.interpret_non_sustained_ventricular_tachycardia()}
        - Paroxysmal supraventricular tachyarrhythmia: {self.interpret_paroxysmal_supraventricular_tachyarrhythmia()}
        - Bradycardia: {self.interpret_bradycardia()}
            """
            
# Generate ECG impressions for all patients
df['ECG_impressions'] = df.apply(lambda row: ECGReport(
    row["Ventricular Extrasystole"], row["Ventricular Tachycardia"], row["Non-sustained ventricular tachycardia (CH>10)"], row["Paroxysmal supraventricular tachyarrhythmia"], row["Bradycardia"]).generate_report(), axis = 1)

# Test dictionary 
def generate_dictionary(row):
    # Create a dictionary to store non-missing values
    patient_data = {col: row[col] for col in df.columns if pd.notna(row[col])}
    return patient_data
generate_dictionary(df.iloc[0])

# Test prompt 
def generate_prompt(row):
    # Create a dictionary to store non-missing values
    patient_data = {col: row[col] for col in df.columns if pd.notna(row[col])}

    # Start the prompt
    # prompt = "Generate a structured clinical note based on the following data:\n\n"
    prompt = ""
    
    # Add demographic information 
    if "Age" in patient_data:
        prompt += f"Age: {patient_data['Age']}\n"
    if "Gender (male=1)" in patient_data:
        if patient_data['Gender (male=1)'] == 1:
            prompt += f"Gender: Male \n"
        elif patient_data['Gender (male=1)'] == 0:
            prompt += f"Gender: Female \n"
    if "Weight (kg)" in patient_data:
        prompt += f"Weight: {patient_data['Weight (kg)']} kg\n"
    if "Height (cm)" in patient_data:
        prompt += f"Height: {patient_data['Height (cm)']} cm\n"

    # Add clinical features
    if "NYHA class" in patient_data:
        if patient_data['NYHA class'] == 2:
            prompt += f"NYHA Class: II\n"
        elif patient_data['NYHA class'] == 3:
            prompt += f"NYHA Class: III\n"
    if ("Systolic blood pressure (mmHg)" in patient_data) and ("Diastolic blood  pressure (mmHg)" in patient_data):
        prompt += f"Blood Pressure: {patient_data['Systolic blood pressure (mmHg)']}/{patient_data['Diastolic blood  pressure (mmHg)']} mmHg\n"

    # Past medical history
    past_medical_conditions = []
    for condition in ["HF etiology - Diagnosis", "Diabetes (yes=1)", "History of dyslipemia (yes=1)", "Peripheral vascular disease (yes=1)",
                     "History of hypertension (yes=1)", "Prior Myocardial Infarction (yes=1)"]:
        if (condition in patient_data) and (condition == "HF etiology - Diagnosis"):
            if patient_data['HF etiology - Diagnosis'] == 1:
                past_medical_conditions.append("Idiopathic dilated cardiomyopathy")
                # prompt += f"HF Etiology: Idiopathic dilated cardiomyopathy\n"
            if patient_data['HF etiology - Diagnosis'] == 2:
                past_medical_conditions.append("Ischemic dilated cardiomyopathy")
                # prompt += f"HF Etiology: Ischemic dilated cardiomyopathy\n"
            if patient_data['HF etiology - Diagnosis'] == 3:
                past_medical_conditions.append("Enolic dilated cardiomyopathy")
                # prompt += f"HF Etiology: Enolic dilated cardiomyopathy\n"
            if patient_data['HF etiology - Diagnosis'] == 4:
                past_medical_conditions.append("Valvular cardiomyopathy")
                # prompt += f"HF Etiology: Valvular cardiomyopathy\n"
            if patient_data['HF etiology - Diagnosis'] == 5:
                past_medical_conditions.append("Toxic dilated cardiomyopathy")
                #prompt += f"HF Etiology: Toxic dilated cardiomyopathy\n"
            if patient_data['HF etiology - Diagnosis'] == 6:
                past_medical_conditions.append("Post-myocardial dilated cardiomyopathy")
                # prompt += f"HF Etiology: Post-myocardial dilated cardiomyopathy\n"
            if patient_data['HF etiology - Diagnosis'] == 7:
                past_medical_conditions.append("Hypertropic cardiomyopathy")
                # prompt += f"HF Etiology: Hypertropic cardiomyopathy\n"
            if patient_data['HF etiology - Diagnosis'] == 8:
                past_medical_conditions.append("Hypertensive cardiomyopathy")
                # prompt += f"HF Etiology: Hypertensive cardiomyopathy\n"
            if patient_data['HF etiology - Diagnosis'] == 9:
                past_medical_conditions.append("Other HF etiology")
                # prompt += f"HF Etiology: Other\n"
        elif (condition in patient_data) and (condition == "Diabetes (yes=1)"):
            if patient_data['Diabetes (yes=1)'] == 1:
                past_medical_conditions.append("Diabetes")
        elif (condition in patient_data) and (condition == "History of dyslipemia (yes=1)"):
            if patient_data['History of dyslipemia (yes=1)'] == 1:
                past_medical_conditions.append("Dyslipemia")
        elif (condition in patient_data) and (condition == "Peripheral vascular disease (yes=1)"):
            if patient_data['Peripheral vascular disease (yes=1)'] == 1:
                past_medical_conditions.append("Peripheral vascular disease")
        elif (condition in patient_data) and (condition == "History of hypertension (yes=1)"):
            if patient_data['History of hypertension (yes=1)'] == 1:
                past_medical_conditions.append("Hypertension")
        elif (condition in patient_data) and (condition == "Prior Myocardial Infarction (yes=1)"):
            if patient_data['Prior Myocardial Infarction (yes=1)'] == 1:
                past_medical_conditions.append("Myocardial Infarction")
    if past_medical_conditions:
        prompt += "Past Medical History: " + ", ".join(past_medical_conditions) + "\n"
    else:
        prompt += "Past Medical History: None reported.\n"

    # Lab results
    lab_tests = ['Albumin (g/L)', 'ALT or GPT (IU/L)', 'AST or GOT (IU/L)', 'Total Cholesterol (mmol/L)', 'Creatinine (?mol/L)',
                 'Gamma-glutamil transpeptidase (IU/L)', 'Glucose (mmol/L)', 'Hemoglobin (g/L)', 'HDL (mmol/L)', 
                 'Potassium (mEq/L)', 'LDL (mmol/L)', 'Sodium (mEq/L)', 'Pro-BNP (ng/L)',  'Protein (g/L)', 'T3 (pg/dL)', 
                 'T4 (ng/L)', 'Troponin (ng/mL)', 'TSH (mIU/L)', 'Urea (mg/dL)']
    for test in lab_tests:
        # Special case for units of creatining
        if (test in patient_data) and (test == 'Creatinine (?mol/L)'):
            prompt += f"Creatinine (mmol/L): {patient_data[test]}\n"
        elif test in patient_data:
            prompt += f"{test}: {patient_data[test]}\n"
    
    # LVEF
    if "LVEF (%)" in patient_data:
        # prompt += f"LVEF (%): {patient_data["LVEF (%)"]}\n"
        prompt += f"LVEF (%): {patient_data['LVEF (%)']}\n"
        
    # Medication
    current_medications = []
    medications = {
    'Calcium channel blocker (yes=1)': "Calcium Channel Blocker",
    'Diabetes medication (yes=1)': "Diabetes Medication",
    'Amiodarone (yes=1)': "Amiodarone",
    'Angiotensin-II receptor blocker (yes=1)': "Angiotensin II Receptor Blocker",
    'Anticoagulants/antitrombotics (yes=1)': "Anticoagulants/Antithrombotics",
    'Betablockers (yes=1)': "Beta Blockers",
    'Digoxin (yes=1)': "Digoxin",
    'Loop diuretics (yes=1)': "Loop Diuretics",
    'Spironolactone (yes=1)': "Spironolactone",
    'Statins (yes=1)': "Statins",
    'Hidralazina (yes=1)': "Hydralazine",
    'ACE inhibitor (yes=1)': "ACE Inhibitor",
    'Nitrovasodilator (yes=1)': "Nitrovasodilator"
    }
    for key, value in medications.items():
        if (key in patient_data) and (patient_data[key]) == 1:
            current_medications.append(value)
    if current_medications:
        prompt += "Medications: " + ", ".join(current_medications) + "\n"
    else:
        prompt += "Medications: None reported.\n"

    # Holter ECG rhythm
    if "Holter  rhythm" in patient_data:
        if patient_data["Holter  rhythm"] == 0:
            prompt += "Holter rhythm: sinus \n"
        elif patient_data["Holter  rhythm"] == 1:
            prompt += "Holter rhythm: permanent atrial fibrillation \n"
        elif patient_data["Holter  rhythm"] == 2:
            prompt += "Holter rhythm: atrial flutter \n"
        elif patient_data["Holter  rhythm"] == 3:
            prompt += "Holter rhythm: pacemaker \n"
    
    # Holter ECG Features (Impressions only)
    prompt += patient_data['ECG_impressions']
    return prompt
    
print(generate_prompt(df.iloc[1]))

# Create prompt dataframe
df_prompts = df[[df.columns[0]]]
df_prompts['Prompts'] = None
for i in range(len(df_prompts)):
    df_prompts.loc[i, 'Prompts'] = generate_prompt(df.iloc[i])
df_prompts.head()

# Map cause of death
mapping = {0: 'survivor', 3: 'sudden cardiac death', 7: 'pump failure death'}

# Create a new column with mapped values
df_prompts['Outcome'] = df['Cause of death'].map(mapping)
df_prompts.head()

print(df_prompts.iloc[0,1])

# Save results
df_prompts.to_csv("../Data/subject-info-cleaned-with-prompts.csv")

# Get subset of results for testing
n_samples = len(df_prompts) # All prompts

# Perform stratified sampling
df_sampled = df_prompts.groupby('Outcome', group_keys=False).apply(lambda x: x.sample(n=min(len(x), n_samples)))

# Save results
df_sampled.to_csv("../Data/subject-info-cleaned-with-prompts-repeated.csv")

# Generate repeats of each row
n = 10
df_sampled_repeated = df_sampled.loc[df_sampled.index.repeat(n)].reset_index(drop=True)

# Save results
df_sampled_repeated.to_csv("../Data/subject-info-cleaned-with-prompts-repeated.csv")
